chore: moved calculateBallotResults to a separate file
